refactor: nectar missions operations
