+++ packages/hpl-nectar-missions/operations/index.ts
@@ -1,6 +1,3 @@
-export * from "./createMissionPool";
-export * from "./updateMissionPool";
-export * from "./createMission";
-export * from "./updateMission";
-export * from "./participate";
-export * from "./recall";
+export * from "./missionPoolOperations";
+export * from "./missionOperations";
+export * from "./participateOperations";
